refactored model out into submodules. wip input parsing.
