+++ parse/src/binary.rs
@@ -1,1 +1,3 @@
 pub mod game;
+
+const EXT_BIN: &'static str = ".bin";
